// src/lib/api/generated/model/milestoneResponseDtoStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * BEACON Work Progress API
 * Secure client dashboard API - projects, milestones, updates, feedback.
 * OpenAPI spec version: 1.0
 */

export type MilestoneResponseDtoStatus = typeof MilestoneResponseDtoStatus[keyof typeof MilestoneResponseDtoStatus];


export const MilestoneResponseDtoStatus = {
  PENDING: 'PENDING',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
} as const;