// src/lib/api/generated/model/milestoneResponseDtoStage.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * BEACON Work Progress API
 * Secure client dashboard API - projects, milestones, updates, feedback.
 * OpenAPI spec version: 1.0
 */

export type MilestoneResponseDtoStage = typeof MilestoneResponseDtoStage[keyof typeof MilestoneResponseDtoStage];


export const MilestoneResponseDtoStage = {
  DISCOVER: 'DISCOVER',
  DESIGN: 'DESIGN',
  BUILD: 'BUILD',
  SCALE: 'SCALE',
} as const;