// src/lib/api/generated/model/projectResponseDtoStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * BEACON Work Progress API
 * Secure client dashboard API - projects, milestones, updates, feedback.
 * OpenAPI spec version: 1.0
 */

export type ProjectResponseDtoStatus = typeof ProjectResponseDtoStatus[keyof typeof ProjectResponseDtoStatus];


export const ProjectResponseDtoStatus = {
  ACTIVE: 'ACTIVE',
  ON_HOLD: 'ON_HOLD',
  COMPLETED: 'COMPLETED',
} as const;